refactor(tests): give probe_endpoints its own httpx test file
Folding the probe_endpoints parse/stdin/missing-binary cases into
httpx/test_scanner.py pushed it to 461 lines, over the 450 bar. The shim
is a distinct public entry point (httpx_scan in TECH_DETECT mode), so it
earns its own test_probe_endpoints.py - test_scanner.py drops back to 371
and stays focused on scanner.py internals. Net-zero on tests.
